simplified downsampling and moved it to ymltools
